Move common files to core-ui
