+++ program-ui/nft/nft-drop-panel.tsx
@@ -4,8 +4,8 @@ import { useLazyMint } from "@thirdweb-dev/react/solana";
 import { NFTMetadataInput } from "@thirdweb-dev/sdk";
 import type { NFTDrop } from "@thirdweb-dev/sdk/solana";
 import { UploadProgressEvent } from "@thirdweb-dev/storage";
-import { BatchLazyMint } from "contract-ui/tabs/nfts/components/batch-lazy-mint";
 import { NFTMintForm } from "contract-ui/tabs/nfts/components/mint-form";
+import { BatchLazyMint } from "core-ui/batch-upload/batch-lazy-mint";
 import { useTrack } from "hooks/analytics/useTrack";
 import { useTxNotifications } from "hooks/useTxNotifications";
 import { useState } from "react";
